Enhance voice output and update project dependencies
Improved the user experience by enhancing the voice output quality and adding a more engaging tone. Adjusted voice modulation and timing for a more natural and friendly speech. Minor tweaks to the script for clarity and better alignment with the intended message. Updated the audio file to reflect these changes.

Updated `TypeCatch.Net5.csproj` to target .NET 9.0 and added `SupportedOSPlatformVersion` property with a value of `8.0`. Upgraded `EntityFramework` to `6.5.1`, `Microsoft.Extensions.Configuration.UserSecrets` to `8.0.0`, and `NAudio` to `2.2.1`.

In `MainVM.Core.cs`, added a condition to check if `TodoToday` is less than or equal to 0, canceling all ongoing speech synthesis and announcing "You are off the hook for today." Modified several motivational messages by removing emojis and adjusting the text.

Updated `WindowsFormsLib.csproj` to target .NET 9.0 and changed `SupportedOSPlatformVersion` to `7.0`. Removed the project reference to `TypeCatch.Net5.csproj`.